### 2024-03-20 #6508 @danielskatz
👋 @Huzaifg - thanks for your submission to JOSS.

As we get started in the process of finding an editor and moving to review, you could take a look at the generated PDF, as I see a few minor things that could be fixed (text that says Fig. Figure), a reference in the text that has extra parenthesis (see how references are cited in our [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)), etc.

Also, you could work on the possibly missing DOI that editorialbot suggests, and check on the ones where there might or might not be a DOI, but note that some may be incorrect.  Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again. You can also use the command `@editorialbot generate pdf` when the references are right or you've made changes in the .md file to make a new PDF. editorialbot commands need to be the first entry in a new comment.

A couple of other small things I notices when browsing the repo:
* The README in https://github.com/uwsbel/low-fidelity-dynamic-models/tree/main/wheeled_vehicle_models/18dof talks about the 11 DOF model in the first line, which I think should be the 18 DOF model.
* Lines 110 and 152 of https://github.com/uwsbel/low-fidelity-dynamic-models/blob/main/wheeled_vehicle_models/18dof/dof18.cpp have "rigth" instead of "right"

These imply to me that you created one model/set of code initially, then made copies and changed them for the other models and other parts of the models, but this of course can lead to copy and paste errors. It would be great if someone else on your team could do a careful check of the code/repo before the review starts.

Given all of this, I'm going to mark this as paused for now, but please let me know when you've completed these things, then we'll re-start the process at that point.
