### 2024-03-22 8688c75
all figs and ref paranthesis fixed. Only DOI for thesis remains
