+++ paper/paper.bib
@@ -1,16 +1,15 @@
-@inproceedings{Rill:2013,
-  author = {Rill, Georg},
-  year   = {2013},
-  month  = {01},
-  pages  = {},
-  title  = {{TM}easy -- A handling tire model based on a three-dimensional slip approach}
-}
-
-@misc{rapidJSON:2017,
-  author       = {Milo Yip},
-  title        = {{RapidJSON}: A fast {JSON} parser/generator for C++ with both {SAX/DOM} style {API}},
-  howpublished = {\url{http://rapidjson.org/index.html}},
-  note         = {Accessed: 2017-10-20}
+@article{Hirschberg:2002,
+  author    = {Wolfgang Hirschberg and Georg Rill and Heinz Weinfurter},
+  title     = {User-Appropriate Tyre-Modelling for Vehicle Dynamics in Standard and Limit Situations},
+  journal   = {Vehicle System Dynamics},
+  volume    = {38},
+  number    = {2},
+  pages     = {103-125},
+  year      = {2002},
+  publisher = {Taylor \& Francis},
+  doi       = {10.1076/vesd.38.2.103.5620},
+  URL       = {https://www.tandfonline.com/doi/abs/10.1076/vesd.38.2.103.5620},
+  eprint    = {https://www.tandfonline.com/doi/pdf/10.1076/vesd.38.2.103.5620}
 }
 
 @article{Serban:2019,
@@ -18,9 +17,11 @@
   author  = {Serban, R. and Taylor, M. and Negrut, D. and Tasora, A.},
   journal = {Intl. J. Veh. Performance},
   year    = {2019},
+  month = {01},
   volume  = {5},
   pages   = {18--39},
-  number  = {1}
+  number  = {1},
+  doi = {10.1504/IJVP.2019.10018132}
 }
 
 @phdthesis{huzaifa:2023,
@@ -80,11 +81,16 @@
   eprint    = {https://doi.org/10.1080/00423114.2016.1223863}
 }
 
-@inproceedings{Ascher:1998,
-  title  = {Computer methods for ordinary differential equations and differential-algebraic equations},
-  author = {Uri M. Ascher and Linda R. Petzold},
-  year   = {1998},
-  url    = {https://api.semanticscholar.org/CorpusID:32366732}
+@book{Ascher:1998,
+author = {Ascher, Uri M. and Petzold, Linda R.},
+title = {Computer Methods for Ordinary Differential Equations and Differential-Algebraic Equations},
+publisher = {Society for Industrial and Applied Mathematics},
+year = {1998},
+doi = {10.1137/1.9781611971392},
+address = {Philadelphia, PA},
+edition   = {},
+URL = {https://epubs.siam.org/doi/abs/10.1137/1.9781611971392},
+eprint = {https://epubs.siam.org/doi/pdf/10.1137/1.9781611971392}
 }
 
 @article{Hindmarsh:2005,
