### 2024-03-22 4286155
ensure consistency for DOF and corrected spelling
