+++ wheeled_vehicle_models/18dof/README.md
@@ -1,5 +1,5 @@
 ## Overview
-The 18 Degree Of Freedom (DOF) vehicle model is a double-track vehicle model with simplified roll and no pitch. This is the intermediate level of fidelity and is thus faster than the 24dof model but slower than the 11 DOF model. The 18 DOF model does not capture pitch and heave motions, and the front and rear suspension are represented simply by their respective roll stiffness and roll damping coefficients. Additionally, wheel lift off conditions cannot be simulated. The vertical forces $F_{z_{ij}}$ are obtained based on quasi-static lateral and longitudinal load transfers. This model assumes the same engine, torque converter, powertrain and TMeasy tires as the 11 DOF and 24 DOF models. The 18 DOF model takes the same input as the 11 DOF and 24 DOF models - Throttle $\in [0,1]$, Braking $\in [0,1]$ and Steering $\in [-1,1]$ where $-1$ is a full left turn. See chapter 2 [here](https://uwmadison.box.com/s/2tsvr4adbrzklle30z0twpu2nlzvlayc) for more details.
+The 18 Degree Of Freedom (DOF) vehicle model is a double-track vehicle model with simplified roll and no pitch. This is the intermediate level of fidelity and is thus faster than the 24 DOF model but slower than the 11 DOF model. The 18 DOF model does not capture pitch and heave motions, and the front and rear suspension are represented simply by their respective roll stiffness and roll damping coefficients. Additionally, wheel lift off conditions cannot be simulated. The vertical forces $F_{z_{ij}}$ are obtained based on quasi-static lateral and longitudinal load transfers. This model assumes the same engine, torque converter, powertrain and TMeasy tires as the 11 DOF and 24 DOF models. The 18 DOF model takes the same input as the 11 DOF and 24 DOF models - Throttle $\in [0,1]$, Braking $\in [0,1]$ and Steering $\in [-1,1]$ where $-1$ is a full left turn. See chapter 2 [here](https://uwmadison.box.com/s/2tsvr4adbrzklle30z0twpu2nlzvlayc) for more details.
 
 ### Where do I set model parameters?
 Model parameters are set through JSON files which are then used in the `construct` solver function. Examples of these JSON files are available for 2 different vehicles in [data/json/HMMWV](./data/json/HMMWV) and [data/json/Sedan](./data/json/Sedan). Examples of use are available in [demos](./demos/).
@@ -39,7 +39,7 @@ Once Sundials is built successfully, `USE_SUNDIALS` is set to `ON` and `c` is hi
 - `SUNDIALS_DIR` - Here, set the absolute path to the sundials build directory. This is the directory in which the `SUNDIALSConfig.cmake` file is found.
 Once this path is correctly set, hit `c` to configure.
 
-Once all the required options are set hit `c` to configure and `g` to generate. This will then populate your `build` directory with the necessary `makefile` with which you can build the 18dof model along with its demos by running
+Once all the required options are set hit `c` to configure and `g` to generate. This will then populate your `build` directory with the necessary `makefile` with which you can build the 18 DOF model along with its demos by running
 ```bash
 make -j4
 ```
@@ -47,7 +47,7 @@ We recommend to not run `make install` to ensure path consistencies for the demo
 Once this is run, the `build` folder will be populated with the executables of the [demos](./demos) and the Python library (if `BUILD_PYTHON_MODULE` is set to `ON`).  
 
 ### Python Wrapper
-Since the Python module is built using SWIG, the path to the Python wrapper library needs to be appened to `PYTHON_PATH`. This library can be found in the build directory mentioned above and its path can be set (on Linux) permanently using the following instructions
[…]
