+++ paper/paper.bib
@@ -29,8 +29,9 @@
   title     = {A set of reduced order vehicle models for applications in machine learning, state estimation and control},
   school    = {University of Wisconsin--Madison},
   year      = {2023},
-  type      = {{M.S.} thesis, \url{https://uwmadison.box.com/s/zfy7zhg7c6i7nexfoo2r2r6fop5cu9it}},
-  timestamp = {2023.05.02}
+  type      = {M.S. thesis},
+  url = {https://minds.wisconsin.edu/handle/1793/85062},
+  doi = {1793/85062}
 }
 
 @inproceedings{Pepy:2006,
