+++ wheeled_vehicle_models/24dof/dof24.h
@@ -130,8 +130,8 @@ struct TMeasyParam {
     // Longitudinal
     double _dfx0Pn;   //!< Initial longitudinal slopes dFx/dsx [N] at Nominal load
     double _dfx0P2n;  //!< Intial longitudinal slopes dFx/dsx [N] at max load
-    double _fxmPn;    //!< Maximum longituidnal force [N] at nominal load
-    double _fxmP2n;   //!< Maximum longituidnal force [N] at max load
+    double _fxmPn;    //!< Maximum longitudinal force [N] at nominal load
+    double _fxmP2n;   //!< Maximum longitudinal force [N] at max load
     double _fxsPn;    //!< Longitudinal load at sliding [N] at nominal load
     double _fxsP2n;   //!< Longitudinal load at sliding [N] at max load
     double _sxmPn;    //!< Slip sx at maximum longitudinal load Fx at nominal load
@@ -187,7 +187,7 @@ struct TMeasyState {
 
     /// @brief Constructor that takes in all the states and sets them to the values that are passed in. Note, its
     /// usually very difficult to set all the states to physical meaningful values while ensuring that the tire is in
-    /// equilibrium. This is also largely untested and its thus recommended to initiaize the states to 0 or use the copy
+    /// equilibrium. This is also largely untested and its thus recommended to initialize the states to 0 or use the copy
     /// constructor below
     TMeasyState(double xe,
                 double ye,
@@ -410,8 +410,8 @@ struct TMeasyNrParam {
     // Longitudinal
     double _dfx0Pn;   //!< Initial longitudinal slopes dFx/dsx [N] at Nominal load
     double _dfx0P2n;  //!< Intial longitudinal slopes dFx/dsx [N] at max load
-    double _fxmPn;    //!< Maximum longituidnal force [N] at nominal load
-    double _fxmP2n;   //!< Maximum longituidnal force [N] at max load
+    double _fxmPn;    //!< Maximum longitudinal force [N] at nominal load
+    double _fxmP2n;   //!< Maximum longitudinal force [N] at max load
     double _fxsPn;    //!< Longitudinal load at sliding [N] at nominal load
     double _fxsP2n;   //!< Longitudinal load at sliding [N] at max load
     double _sxmPn;    //!< Slip sx at maximum longitudinal load Fx at nominal load
@@ -539,7 +539,7 @@ struct TMeasyNrState {
           _My(other._My),
           _engTor(other._engTor) {}
 
-    // Sates that are integrated - these are actually wheel states but we don't use that abstraction
+    // States that are integrated - these are actually wheel states but we don't use that abstraction
     double _omega;   //!< Angular velocity of wheel
     double _dOmega;  //!< Angular acceleration of wheel
     double _xt;      //!< Vertical tire compression
@@ -637,7 +637,7 @@ struct VehicleParam {
           _CFmap(other._CFmap),
           _TRmap(other._TRmap) {}
 
-    // Declaration of all the 14 DOF vehicle parameters pretty much the same as the 8DOF model
+    // Declaration of all the 24 DOF vehicle parameters pretty much the same as the 18 DOF model
     double _a;     //!< Distance of C.G to front axle (m)
     double _b;     //!< Distance of C.G to rear axle (m)
     double _h;     //!< Height of C.G
@@ -690,7 +690,7 @@ struct VehicleParam {
     bool _tcbool;  //!< Boolean that checks for the presence of a torque converter. Can be set using "tcBool" in the
                    //!< JSON file. Defaults to 0 if not specified.
 
-    double _maxBrakeTorque;  //!< The maximum braking torque (Nm) that the brakes applu to the wheels. Can be set using
+    double _maxBrakeTorque;  //!< The maximum braking torque (Nm) that the brakes apply to the wheels. Can be set using
[…]
