### 2024-03-25 #6508 @Huzaifg
Hello @danielskatz!

PR [#16](https://github.com/uwsbel/low-fidelity-dynamic-models/pull/16) and [#17](https://github.com/uwsbel/low-fidelity-dynamic-models/pull/17) resolve all the issues raised. I think we are ready to restart the process
