### 2024-03-26 #6508 @danielskatz
👋 @cst0 and @joeybernard - Would you be willing to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
