### 2024-03-27 #6548 @editorialbot
[REVIEW]: A Library of Lower Fidelity Dynamics Models (LFDMs) For On-Road Vehicle Dynamics Targeting Faster Than Real-Time Applications
**Submitting author:** @Huzaifg (<a href="http://orcid.org/0009-0004-4273-1212">HUZAIFA MUSTAFA UNJHAWALA</a>)
**Repository:** https://github.com/uwsbel/low-fidelity-dynamic-models
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
