### 2024-04-12 uwsbel/low-fidelity-dynamic-models#18 @joeybernard
which C++ compiler is being used
I tried building the code from the repo. I ran into errors when trying to compile the code for 18dof.  Using g++, I get the following error:
--------------------------------------------
Consolidate compiler generated dependencies of target demo_hmmwv_hi
[  3%] Building CXX object CMakeFiles/demo_hmmwv_hi.dir/dof18.cpp.o
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:296:53: error: ‘void d18::computeCombinedcoulombForce(double&, double&, double, double, double, double, double)’ should have been declared inside ‘d18’
  296 |                                      double vcoulomb) {
      |                                                     ^
make[2]: *** [CMakeFiles/demo_hmmwv_hi.dir/build.make:90: CMakeFiles/demo_hmmwv_hi.dir/dof18.cpp.o] Error 1
make[1]: *** [CMakeFiles/Makefile2:93: CMakeFiles/demo_hmmwv_hi.dir/all] Error 2
make: *** [Makefile:91: all] Error 2
--------------------------------------------

Moving to using clang++, version 14, I get the following error:
--------------------------------------------
[  6%] Building CXX object CMakeFiles/demo_hmmwv_hi.dir/dof18.cpp.o
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:290:11: error: out-of-line definition of 'computeCombinedcoulombForce' does not match any declaration in namespace 'd18'; did you mean 'computeCombinedColumbForce'?
void d18::computeCombinedcoulombForce(double& fx,
          ^~~~~~~~~~~~~~~~~~~~~~~~~~~
          computeCombinedColumbForce
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.h:894:6: note: 'computeCombinedColumbForce' declared here
void computeCombinedColumbForce(double& fx, double& fy, double mu, double vsx, double vsy, double fz, double vcoulomb);
     ^
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:593:5: error: use of undeclared identifier 'computeCombinedcoulombForce'; did you mean 'computeCombinedColumbForce'?
    computeCombinedcoulombForce(Fx0, Fy0, t_params._mu, vsx, vsy, fz, t_params._vcoulomb);
    ^~~~~~~~~~~~~~~~~~~~~~~~~~~
    computeCombinedColumbForce
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:290:11: note: 'computeCombinedColumbForce' declared here
void d18::computeCombinedcoulombForce(double& fx,
          ^
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:749:13: warning: taking the absolute value of unsigned type 'bool' has no effect [-Wabsolute-value]
        if (std::abs((v_states._u - 0) < 1e-9) && (torque_t < 0)) {
            ^
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:749:13: note: remove the call to 'abs' since unsigned values cannot be negative
        if (std::abs((v_states._u - 0) < 1e-9) && (torque_t < 0)) {
            ^~~~~~~~
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:783:13: warning: taking the absolute value of unsigned type 'bool' has no effect [-Wabsolute-value]
        if (std::abs((v_states._u - 0) < 1e-9) && (torque_t < 0)) {
            ^
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:783:13: note: remove the call to 'abs' since unsigned values cannot be negative
        if (std::abs((v_states._u - 0) < 1e-9) && (torque_t < 0)) {
            ^~~~~~~~
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:921:13: warning: taking the absolute value of unsigned type 'bool' has no effect [-Wabsolute-value]
        if (std::abs((v_states._u - 0) < 1e-9) && (torque_t < 0)) {
            ^
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:921:13: note: remove the call to 'abs' since unsigned values cannot be negative
        if (std::abs((v_states._u - 0) < 1e-9) && (torque_t < 0)) {
            ^~~~~~~~
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:955:13: warning: taking the absolute value of unsigned type 'bool' has no effect [-Wabsolute-value]
        if (std::abs((v_states._u - 0) < 1e-9) && (torque_t < 0)) {
            ^
/home/jbernard/repos/low-fidelity-dynamic-models/wheeled_vehicle_models/18dof/dof18.cpp:955:13: note: remove the call to 'abs' since unsigned values cannot be negative
        if (std::abs((v_states._u - 0) < 1e-9) && (torque_t < 0)) {
            ^~~~~~~~
4 warnings and 2 errors generated.
make[2]: *** [CMakeFiles/demo_hmmwv_hi.dir/build.make:90: CMakeFiles/demo_hmmwv_hi.dir/dof18.cpp.o] Error 1
make[1]: *** [CMakeFiles/Makefile2:93: CMakeFiles/demo_hmmwv_hi.dir/all] Error 2
make: *** [Makefile:91: all] Error 2
--------------------------------------------
