### 2024-04-12 #6548 @joeybernard
I ran into a compiling error. I created issue https://github.com/uwsbel/low-fidelity-dynamic-models/issues/18
