+++ wheeled_vehicle_models/18dof/dof18.cpp
@@ -287,7 +287,7 @@ void d18::tmxy_combined(double& f, double& fos, double s, double df0, double sm,
     }
 }
 
-void d18::computeCombinedcoulombForce(double& fx,
+void d18::computeCombinedCoulombForce(double& fx,
                                      double& fy,
                                      double mu,
                                      double vsx,
@@ -590,7 +590,7 @@ void d18::computeTireRHS(TMeasyNrState& t_states,
     // Compute the combined column force (used for low speed stability)
     double Fx0 = 0;
     double Fy0 = 0;
-    computeCombinedcoulombForce(Fx0, Fy0, t_params._mu, vsx, vsy, fz, t_params._vcoulomb);
+    computeCombinedCoulombForce(Fx0, Fy0, t_params._mu, vsx, vsy, fz, t_params._vcoulomb);
 
     // evaluate the slips
     double sx = -vsx / vta;
