+++ wheeled_vehicle_models/18dof-gpu/dof18_gpu.cu
@@ -284,7 +284,7 @@ d18GPU::tmxy_combined(double* f, double* fos, double s, double df0, double sm, d
     }
 }
 
-__device__ void d18GPU::computeCombinedcoulombForce(double* fx,
+__device__ void d18GPU::computeCombinedCoulombForce(double* fx,
                                                    double* fy,
                                                    double mu,
                                                    double vsx,
@@ -582,7 +582,7 @@ __device__ void d18GPU::computeTireRHS(TMeasyNrState* t_states,
     // Compute the combined column force (used for low speed stability)
     double Fx0 = 0;
     double Fy0 = 0;
-    computeCombinedcoulombForce(&Fx0, &Fy0, t_params->_mu, vsx, vsy, fz, t_params->_vcoulomb);
+    computeCombinedCoulombForce(&Fx0, &Fy0, t_params->_mu, vsx, vsy, fz, t_params->_vcoulomb);
 
     // evaluate the slips
     double sx = -vsx / vta;
