+++ wheeled_vehicle_models/24dof-gpu/dof24_gpu.cu
@@ -817,7 +817,7 @@ d24GPU::tmxy_combined(double* f, double* fos, double s, double df0, double sm, d
     }
 }
 
-__device__ void d24GPU::computeCombinedColumbForce(double* fx,
+__device__ void d24GPU::computeCombinedCoulombForce(double* fx,
                                                    double* fy,
                                                    double mu,
                                                    double vsx,
@@ -1041,7 +1041,7 @@ __device__ void d24GPU::computeTireRHS(const VehicleState* v_states,
     // Compute the combined column force (used for low speed stability)
     double Fx0 = 0;
     double Fy0 = 0;
-    computeCombinedColumbForce(&Fx0, &Fy0, t_params->_mu, vsx, vsy, fz, t_params->_vcoulomb);
+    computeCombinedCoulombForce(&Fx0, &Fy0, t_params->_mu, vsx, vsy, fz, t_params->_vcoulomb);
 
     // evaluate the slips
     double sx = -vsx / vta;
