+++ wheeled_vehicle_models/11dof/dof11.cpp
@@ -208,7 +208,7 @@ void d11::tmxy_combined(double& f, double& fos, double s, double df0, double sm,
         fos = 0.0;
     }
 }
-void d11::computeCombinedColumbForce(double& fx,
+void d11::computeCombinedCoulombForce(double& fx,
                                      double& fy,
                                      double mu,
                                      double vsx,
@@ -522,7 +522,7 @@ void d11::computeTireRHS(TMeasyNrState& t_states,
     // Compute the combined column force (used for low speed stability)
     double Fx0 = 0;
     double Fy0 = 0;
-    computeCombinedColumbForce(Fx0, Fy0, t_params._mu, vsx, vsy, fz, t_params._vcoulomb);
+    computeCombinedCoulombForce(Fx0, Fy0, t_params._mu, vsx, vsy, fz, t_params._vcoulomb);
 
     // evaluate the slips
     double sx = -vsx / vta;
