+++ wheeled_vehicle_models/24dof/dof24_sundials.h
@@ -83,7 +83,7 @@ class UserData {
 
     void SetCurrentGear(int gear) { m_current_gr = gear; }
     int GetCurrentGear() const { return m_current_gr; }
-    TireType GetTireType() const { return m_tire_type; }
+    d24::TireType GetTireType() const { return m_tire_type; }
 
     // Indices of vehicle position in state vectors
     int m_vx_idx;
@@ -102,7 +102,7 @@ class UserData {
     // Current transmission gear
     int m_current_gr;
     // Tire type - changes how we pack and unpack the state vectors
-    TireType m_tire_type;
+    d24::TireType m_tire_type;
 
     // Sensitivity parameters
     int m_param_flags;
@@ -148,7 +148,7 @@ class d24SolverSundials {
                    const std::string& tire_params_file,
                    const std::string& sus_params_file,
                    const std::string& driver_inputs_file,
-                   TireType type);
+                   d24::TireType type);
 
     /// @brief Set the absolute and relative tolerances for the solver.
 
@@ -255,7 +255,7 @@ class d24SolverSundials {
 
     /// @brief Get the TireType
     /// @return TireType (Either TMEasy or TMEasyNr)
-    TireType GetTireType() const { return m_tire_type; }
+    d24::TireType GetTireType() const { return m_tire_type; }
 
     /// @brief Solve the system of equations and run the simulation uptil m_tend
 
@@ -306,7 +306,7 @@ class d24SolverSundials {
     bool m_verbose;             // verbose output?
     bool m_output;              // generate output file?
     std::string m_output_file;  // output file name
-    TireType m_tire_type;       // Tire type - changes how we pack and unpack the state vectors
+    d24::TireType m_tire_type;       // Tire type - changes how we pack and unpack the state vectors
 };
 #ifndef SWIG
 void packY(const d24::VehicleState& v_states,
