### 2024-04-12 uwsbel/low-fidelity-dynamic-models#18 @danielskatz
adding pointer to openjournals/joss-reviews#6548
