### 2024-04-23 #6548 @danielskatz
👋 @joeybernard - It looks like @Huzaifg has addressed your issue in https://github.com/uwsbel/low-fidelity-dynamic-models/issues/18 - Can you make more progress now?
