### 2024-05-02 #6548 @cst0
Overall this is a good contribution with a clear benefit for the open-source community. However, I would recommend some changes before publication. My notes are below.

Some notes from building:

- CMake version 3.12 or 3.18 is not necessary (and is a more future-facing version than is in many widely-used distributions). 3.10 works fine. I would recommend changing the CMake version requirement to a lower value.
- The install instructions mention 'make install' but no such target exists. Running `make` ran fine, and the install target appears to be optional, so I'll proceed.
- There are many subdirectories to be built one-by-one. This is a process that should be automated: one top-level CMakeLists.txt that references each of these subdirectories would be a standard way of doing this.
- I don't see a list of requirements for building/usage. A section in the README at minimum, or (optionally depending on preference) making use of a tool like Conan for c++ package management. In particular, Sundial should be linked, because this is required for building tests.

On replicating:

- It's not clear how to recreate the figures or claims presented in paper.md. Tests are provided, but how are they run? Demos are provided, but how are these used?
- I've found some executables that are generated when building (e.g., `demo_hmmwv`), but these segfault when run directly. Instructions on intended use would be valuable here.

On the written content:

- LaTeX formatting should be checked in `paper.md`. Some other formatting does not display (e.g., paragraph 2 has stray $, some images are only links with {width} tag visible in output).
- Doxygen handles API documentation and appears comprehensive, examples in the paper are well-reasoned and described.
- Overall writing is strong.
