### 2024-05-08 #6548 @joeybernard
The writing in the paper is strong, but there are several formatting typos. Also, should the references appear in the paper itself, rather than just in the separate bibliography file?

In order to reproduce the paper's results, you need to go to the paper/paper_results subdirectory, and the instructions there say to go to the paper_results subdirectory as the first step. This readme should really be up at the top level of the repository, so that users can find these instructions. Running the included bench.sh script compiles the CPU versions correctly and runs the tests. But, it doesn't remember whether or not you asked for the GPU versions. So, if you ask for only the CPU version, it tries to run the GPU version too and generates errors. When you say you do want to generate the GPU versions, it assumes that you have nvcc installed. I did not notice that this was included in the list of requirements. A different option might be to use OpenCL to manage the GPU part, as it won't be tied specifically to Nvidia hardware.

Going to the individual directories under the wheeled_vehicle_models, the CPU versions build fine. But running them gives an error where it aborts. The GPU versions don't compile because you need to have nvcc installed. If there is no plan to move to a more generic system like OpenCL, it should be documented clearly that the only GPU supported is Nvidia.
