### 2024-05-08 uwsbel/low-fidelity-dynamic-models#20 @Huzaifg
Fixing reviwers comments
Fix comments raised at [6548](https://github.com/openjournals/joss-reviews/issues/6548)
