+++ wheeled_vehicle_models/11dof-gpu/CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.18)
+cmake_minimum_required(VERSION 3.10)
 project(d11dof_gpu LANGUAGES CXX CUDA)
 # -----------------------------------------------------
 set(CMAKE_CXX_STANDARD 17)
