+++ wheeled_vehicle_models/18dof/CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.12)
+cmake_minimum_required(VERSION 3.10)
 project(d18dof)
 # -----------------------------------------------------
 set(CMAKE_CXX_STANDARD 17)
