### 2024-05-14 914bd57
build structure changed. Now only one master CMakeLists to build all the models
