### 2024-05-14 dc449dd
update readmes with more info about demos and new build structure
