### 2024-05-14 12f5988
better user input error checking with more descriptive error messages
