+++ wheeled_vehicle_models/24dof-gpu/demos/HMMWV/demo_hmmwv_step_24gpu.cu
@@ -2,16 +2,15 @@
 // Authors: Huzaifa Unjhawala
 // =============================================================================
 //
-// This demo describes simulating user provided number of HMMWVs (specified with JSON files) on a step-by-step
-// basis. The "step" at which a new control input is provided can be set with SetKernelSimTime as shown here.
-// Since the Half Implicit solver is the only one supported for the GPU models, that is what is used here.
-// When the solver is used in a step-by-step manner, the output is not stored in a file (unlike the CPU models).
-// However, access to the vehicle states every control time step is provided through the GetSimState function.
-// Use ./executable_name <total_number_of_vehicles> <threads_per_block>
+// This demo describes simulating a specified number of HMMWVs (configured with JSON files), all operating on the
+// same driver inputs on the GPU. The Half Implicit solver, designed for GPU execution, is used.
+// The API resembles the CPU version but includes specific GPU settings such as vehicle counts and threads per block.
+// Usage: ./executable_name <total_number_of_vehicles> <threads_per_block>
 //
 // =============================================================================
 #include <cuda.h>
 #include <iostream>
+#include <filesystem>
 #include <random>
 #include <cuda_runtime.h>
 #include <math.h>
@@ -22,40 +21,39 @@
 
 #include "dof24_halfImplicit_gpu.cuh"
 
+namespace fs = std::filesystem;
 using namespace d24GPU;
+
 int main(int argc, char** argv) {
-    // Get total number of vehicles from command line
+    if (argc != 3) {
+        std::cerr << "Usage: " << argv[0] << " <total_number_of_vehicles> <threads_per_block>" << std::endl;
+        return 1;
+    }
+
     unsigned int num_vehicles = std::stoul(argv[1]);
-    // Set the threads per block from command line
     unsigned int threads_per_block = std::stoul(argv[2]);
+
     std::string file_name = "acc3";
-    // Driver inputs -> All vehicles have the same driver inputs
-    std::string driver_file = "../../24dof-gpu/data/input/" + file_name + ".txt";
+    std::string inputPath = "../../24dof-gpu/data/input/" + file_name + ".txt";
 
-    // Vehicle specification -> We assume that all vehicles have the same parameters
-    std::string vehParamsJSON = (char*)"../../24dof-gpu/data/json/HMMWV/vehicle.json";
-    std::string tireParamsJSON = (char*)"../../24dof-gpu/data/json/HMMWV/tmeasy.json";
-    std::string susParamsJSON = (char*)"../../24dof-gpu/data/json/HMMWV/suspension.json";
+    // Verify the existence of the input file
+    if (!fs::exists(inputPath)) {
+        std::cerr << "Error: Input file does not exist: " << inputPath << std::endl;
+        return 1;
+    }
 
-    // Construct the solver
+    std::string vehParamsJSON = "../../24dof-gpu/data/json/HMMWV/vehicle.json";
+    std::string tireParamsJSON = "../../24dof-gpu/data/json/HMMWV/tmeasy.json";
+    std::string susParamsJSON = "../../24dof-gpu/data/json/HMMWV/suspension.json";
+
+    // Construct the solver with GPU-specific settings
     d24SolverHalfImplicitGPU solver(num_vehicles);
-    // The number of vehicles here sets these parameters and inputs for all these vehicles
-    // If there is a need to set different parameters for different vehicles, then the solver
-    // needs to be constructed for each vehicle separately (using the same sovler object)
-    // No driver file
-    solver.Construct(vehParamsJSON, tireParamsJSON, susParamsJSON, num_vehicles, driver_file);
[…]

[…]
