### 2024-05-15 0ff0413
sundials build script along with instructions in README
