+++ README.md
@@ -20,18 +20,21 @@ Using CMake, the user can choose to build the models to execute on the CPU or NV
 
 This library seeks to establish the right balance between speed and accuracy in robotics simulation, especially in the context of vehicle dynamics.
 
+## How do I get started?
+Please see the instructions in [wheeled_vehicle_models](./wheeled_vehicle_models/README.md) for details on how to build and run the models.
+
 ## Repository Structure
 
 Several sub-folders contain each the source code for a vehicle model and the numerical integrator (time stepper) that runs it. CMake configurations have been set up to help with building _demos_, enabling the user to quickly kick off a simulation. The main repo folders are as follows:
 
 - **docs**: Documentation related to the dynamic models and their usage. See [README](./docs/README.md) for instructions to build the docs locally.
 - **wheeled_vehicle_models**: Contains subdirectories for different DOF models for wheeled vehicles:
-  - **11dof**: CPU version for the 11 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/11dof/README.md) for build and use instructions.
-  - **11dof-gpu**: GPU version for the 11 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/11dof-gpu/README.md) for build and use instructions.
-  - **18dof**: CPU version for the 18 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/18dof/README.md) for build and use instructions.
-  - **18dof-gpu**: GPU version for the 18 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/18dof-gpu/README.md) for build and use instructions.
-  - **24dof**: CPU version for the 24 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/24dof/README.md) for build and use instructions.
-  - **24dof-gpu**: GPU version for the 24 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/24dof-gpu/README.md) for build and use instructions.
+  - **11dof**: CPU version for the 11 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/11dof/README.md) for use instructions.
+  - **11dof-gpu**: GPU version for the 11 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/11dof-gpu/README.md) for use instructions.
+  - **18dof**: CPU version for the 18 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/18dof/README.md) for use instructions.
+  - **18dof-gpu**: GPU version for the 18 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/18dof-gpu/README.md) for use instructions.
+  - **24dof**: CPU version for the 24 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/24dof/README.md) for use instructions.
+  - **24dof-gpu**: GPU version for the 24 DOF wheeled vehicle model. See [README](./wheeled_vehicle_models/24dof-gpu/README.md) for use instructions.
   - **third_party/rapidjson**: Third-party libraries and dependencies, specifically `rapidjson` for JSON handling.
   - **utils**: Utility scripts and tools supporting the dynamic models.
 Each vehicle model has its own (although very similar) CMake configuration and build procedure. README files are available in each model sub-directory describing this procedure. 
