### 2024-05-15 1ad445d
better readme for reproducing paper results
