### 2024-05-15 0e7b2ef
mention only support for Nvidia GPUs
