### 2024-05-15 6d420d8
fixing optional GPU build feature in paper results
