### 2024-05-16 #6548 @Huzaifg
Hello @joeybernard @cst0 @danielskatz,
We have taken care of most of the concerns raised in [PR #21](https://github.com/uwsbel/low-fidelity-dynamic-models/pull/21) here.
I will now go one by one and point out the commits that deal with each of your concerns

@cst0
1. CMake Version downgraded in [d56ca14](https://github.com/uwsbel/low-fidelity-dynamic-models/pull/21/commits/d56ca14b51b427e7553080949f341294ab976817)
2. Removed `make install` commands in [dc449dd](https://github.com/uwsbel/low-fidelity-dynamic-models/pull/21/commits/dc449ddf081f5590fb8c14c032166d6b4847bce1)
3. One master `CMakeLists.txt` builds all demos now - see [914bd57](https://github.com/uwsbel/low-fidelity-dynamic-models/pull/21/commits/914bd57b8269de0f4cfe8f42e696ab83fd3a236d)
4. List of requirements added to [README.md](https://github.com/uwsbel/low-fidelity-dynamic-models/blob/main/wheeled_vehicle_models/README.md) in folder with the master `CMakeLists.txt`. Additionally, to handle the Sundials dependency, we now provide a [bash script](https://github.com/uwsbel/low-fidelity-dynamic-models/blob/main/wheeled_vehicle_models/buildSundials.sh) to automate building.
5. New [README's](https://github.com/uwsbel/low-fidelity-dynamic-models/blob/main/paper/README.md) in `paper` folder with instructions to recreate figures from paper
6. Instructions in [README.md](https://github.com/uwsbel/low-fidelity-dynamic-models/blob/main/wheeled_vehicle_models/README.md#how-do-i-use-the-models)  describing how to run the demos. Additionally, all the demos are now much more robust and can catch user errors in input arguments. They also provide good error messages to help the user. See commit [12f5988](https://github.com/uwsbel/low-fidelity-dynamic-models/pull/21/commits/12f598801d276171ad38c491bb64b6e6f07baad3)

Outstanding concerns
1. Latex formatting of paper - I am not able to create the formatting error you mention. I tried locally and also checked the pdf that the bot produced [above](https://github.com/openjournals/joss-reviews/issues/6548#issuecomment-2024148257). Could you please point out where exactly the formatting errors are?

@joeybernard
1. From my understanding the references should be included in the paper. See example paper from the JOSS website [here](https://joss.theoj.org/papers/10.21105/joss.00388)
2. Provided a top level [README.md](https://github.com/uwsbel/low-fidelity-dynamic-models/blob/main/paper/README.md) and fixed the confusion related to the instructions to reproduce the paper results
3. Fixed the automated benchmark script to only run the GPU version if the user asks for it. See commit [6d420d8](https://github.com/uwsbel/low-fidelity-dynamic-models/pull/21/commits/6d420d838fece0564bd1c7c2ec1d9c8700715c92)
4. Mention in the repository [README.md](https://github.com/uwsbel/low-fidelity-dynamic-models/blob/main/README.md#key-features) that the code base only supports Nvidia GPUs
5. The seg fault errors should be fixed now with better error handling and meaningful error messages for all the demos. See commit [12f5988](https://github.com/uwsbel/low-fidelity-dynamic-models/pull/21/commits/12f598801d276171ad38c491bb64b6e6f07baad3)

Outstanding concerns
1. Again, I don't find any formatting errors in the pdf produced by the bot [above](https://github.com/openjournals/joss-reviews/issues/6548#issuecomment-2024148257). Please let me know if I am missing anything and I apologize if I am.
