### 2024-05-23 #6548 @joeybernard
I tried to compile the code in the various "dof" subdirectories and ran into the error where it could not find the file "utils.h". Looking at the history of the CMake files, it looks like the include directory sections (around line 61, 62, 63) changed and it can no longer find the utils subdirectory.
