### 2024-05-24 #6548 @joeybernard
OK, I tried it again. The cmake file fails on trying to find nvcc. In order to get it to compile, I needed to install the cuda toolkit. This occurred even if I only turned on one of the CPU only models. Once I installed the cuda toolkit, it worked correctly.  I think either the documentation should state that it requires the cuda toolkit installed, or the cmake file needs to be slightly edited.
