### 2024-05-24 bcf000b
Removes auto searching for CUDA even if GPU models are not choosen by the user
