### 2024-05-24 #6548 @Huzaifg
@joeybernard Thanks for catching this!

I fixed this in [bcf000b](https://github.com/uwsbel/low-fidelity-dynamic-models/commit/bcf000b73cb40cc15928990b4960c4900facba59). I tested it on a machine without nvcc installed and seems to be working fine on my end.
