### 2024-06-03 #6548 @danielskatz
👋 @cst0 - can you let us know if the responses [above](https://github.com/openjournals/joss-reviews/issues/6548#issuecomment-2115445567) from @Huzaifg satisfy you, and let you complete your reviews, or move in that direction? And if not, what else do you think is needed?
