### 2024-07-08 #6548 @cst0
@danielskatz Thanks for the ping, sorry again to be the holdup here. Yes, I've taken another look and I'm satisfied.

Best,
Chris
