### 2024-07-08 #6548 @danielskatz
@cst0 - can you check off the remaining items on [your checklist](https://github.com/openjournals/joss-reviews/issues/6548#issuecomment-2090523839) in this case?
