### 2024-07-08 #6548 @danielskatz
I'm suggesting changes in https://github.com/uwsbel/low-fidelity-dynamic-models/pull/22 . Please merge this, or let me know what you disagree with.
(Note one set of changes is because JOSS affiliations are just organizational affiliations, not roles. If you want to make further changes to add departments, please go ahead.)
