### 2024-07-09 #6548 @danielskatz
👋 @Huzaifg - can you make sure the proof (above) looks good?

After that, please:
