+++ paper/paper.md
@@ -42,7 +42,7 @@ where $\mathbf{x} \in \mathbb{R}^d$ is the $d$ dimensional state of the vehicle,
 
 The LFDM library contains three dynamic vehicle models, differentiated by their Degrees of Freedom (DoF) counts: 11 DoF, 18 DoF, and 24 DoF. Each can be run on the CPU or an NVIDIA GPU.
 
-11 DoF Model: This is a foundational, single-track model with two wheels, primarily employed in controller design. In the literature, it also known as the bicycle model. It features a rigid chassis with three DoFs at the Center of Mass (CM) – yaw, lateral, and longitudinal. The model includes a kinematic driveline, integrating a map-based engine and torque converter, gearbox, and a differential that enables transfer of torque to the wheels. The vehicle model can use one of the two versions of the TMeasy non-linear tire model [@Hirschberg:2002] for traction-force generation. A steering input is converted to a front wheel angle through a steering map.
+11 DoF Model: This is a foundational, single-track model with two wheels, primarily employed in controller design. In the literature, it is also known as the bicycle model. It features a rigid chassis with three DoFs at the Center of Mass (CM) – yaw, lateral, and longitudinal. The model includes a kinematic driveline, integrating a map-based engine and torque converter, gearbox, and a differential that enables transfer of torque to the wheels. The vehicle model can use one of the two versions of the TMeasy non-linear tire model [@Hirschberg:2002] for traction-force generation. A steering input is converted to a front wheel angle through a steering map.
 
 18 DoF Model: A step above the simple 11 DoF model, this double-track model includes four wheels and introduces a roll DoF, enriching the chassis' dynamics. The model shares the driveline structure with the 11 DoF model but adds front and rear differentials for torque distribution across wheels (see \autoref{fig:driveline}). The steering map in this model simultaneously changes the orientation of the front left and right wheels by the same angle. The other model subsystems are identical to those of the 11 DoF model.
 
@@ -65,7 +65,7 @@ In what follows, we discuss a [Statement of need](#statement-of-need) describing
 # Statement of need
 The open-source vehicle models in LFDM aim to provide accurate vehicle simulations that exceed real-time speeds ideal for state estimation, control, reinforcement learning, and traffic simulation, with or without a human in the loop. Informed by existing models in the literature [@Pepy:2006; @Kong:2015; @Jin:2019], they go beyond the state of the art to introduce additional benefits tailored for faster-than-real-time applications.
 
-1. In the domain literature, the focus is on simplistic single-track models with linear tires or fully kinematic models for their speed, despite a trade-off in accuracy. @Liu:2016 found that while a double-track model with non-linear tires is more accurate, its Real Time Factor (RTF - the time require to simulate one second of vehicle dynamics) of 1 and higher limits its use in Control stacks. The LFDM library, with efficient C++ coding, offers faster-than-real-time (RTF<1) double-track models with accurate non-linear tires and realistic subsystems, including drivelines, engines, and torque converters.
+1. In the domain literature, the focus is on simplistic single-track models with linear tires or fully kinematic models for their speed, despite a trade-off in accuracy. @Liu:2016 found that while a double-track model with non-linear tires is more accurate, its Real Time Factor (RTF - the time required to simulate one second of vehicle dynamics) of 1 and higher limits its use in Control stacks. The LFDM library, with efficient C++ coding, offers faster-than-real-time (RTF<1) double-track models with accurate non-linear tires and realistic subsystems, including drivelines, engines, and torque converters.
 
[…]
