### 2024-07-10 #6548 @Huzaifg
Hello @danielskatz
1) I made some minor edits to the paper after proof reading it
2) v1.0.0
3) Zenodo DOI - 10.5281/zenodo.12703022
