### 2024-07-10 #6548 @danielskatz
@openjournals/dev & @xuanxu - The above error is related to a bib entry that has the DOI field `1793/85062` which looked strange to me too, but I verified that https://doi.org/1793/85062 resolved to the correct item.

Reading the [DOI handbook](https://www.doi.org/doi-handbook/DOI_Handbook_Final.pdf) (section 1.3.2, DOI Name Syntax), I see that the prefix can be a multi-level hierarchy with a `.` used to separate the levels, but I don't see that it has to be multiple levels. In this case, it is not, and Crossref seems ok with this.  However, section 3.2.2 does seem to imply that it should be at least 2 levels, and in fact says that it should contain "one or several full stops".

What do you think we should do in this case?
