### 2024-07-10 #6548 @danielskatz
Also, as a kind-of-related comment on the pattern in our validator, the handbook seems to say that the prefix can have multiple levels, and something with more than two would incorrectly fail our check, I think.
