### 2024-07-10 #6548 @danielskatz
With help from @MartinPaulEve via [social media](https://ravenation.club/@mpe/112762496284273811), I now understand that this is really a handle (http://hdl.handle.net/1793/85062), not a DOI (https://doi.org/1793/85062), though the Crossref resolver also "handles" handles.
