### 2024-07-11 #6548 @danielskatz
@Huzaifg - To fix this, I suggest just removing the internal links, which I don't think are really needed in this short paper, as in https://github.com/uwsbel/low-fidelity-dynamic-models/pull/24. If you want to try to keep them, we can do that as well. My guess is the labels probably need to be fully lower-case.
