### 2024-07-12 #6548 @Huzaifg
@danielskatz merged! I don't see the point in keeping the links as well
